Updated the RunScript method to be more dynamic

+++ src/main/java/dev/selena/luacore/utils/lua/LuaManager.java
@@ -2,24 +2,17 @@ package dev.selena.luacore.utils.lua;
 
 import dev.selena.luacore.LuaCore;
 import dev.selena.luacore.utils.text.LuaMessageUtils;
-import org.bukkit.entity.Player;
-import org.bukkit.event.Event;
 import org.luaj.vm2.Globals;
+import org.luaj.vm2.LuaTable;
 import org.luaj.vm2.LuaValue;
 import org.luaj.vm2.lib.jse.CoerceJavaToLua;
 import org.luaj.vm2.lib.jse.JsePlatform;
 
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
 import java.net.URISyntaxException;
-import java.nio.file.FileAlreadyExistsException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Enumeration;
-import java.util.Objects;
-import java.util.jar.JarEntry;
-import java.util.jar.JarFile;
 
 public class LuaManager {
 
@@ -32,18 +25,22 @@ public class LuaManager {
      * @param function The name of the function you want to run
      * @param scriptPath The path to the folder your script is in
      * @param scriptName The name of the script
-     * @param player Used to parse in a player argument
-     * @param event Used for event specific stuff inside the lua script
+     * @param args List of arguments in the form of {@link LuaArgValue LuaArgValue}
      * @return True if the function exists in the Lua file
      */
-    public static boolean runScript(String function, String scriptPath, String scriptName, Player player, Event event) {
+    public static boolean runScript(String function, String scriptPath, String scriptName, LuaArgValue ... args) {
 
         Long start = System.currentTimeMillis();
         scriptName = scriptName.endsWith(".lua") ? scriptName : scriptName + ".lua";
 
         LuaMessageUtils.verboseMessage("Attempting to run script: " + scriptName);
         // Load the Lua script from file
-        String scriptFilePath = new File(LuaCore.getPlugin().getDataFolder() + "/" + scriptPath, scriptName).getPath();
+        File scriptFile = new File(LuaCore.getPlugin().getDataFolder() + "/" + scriptPath, scriptName);
+        if (!scriptFile.exists()) {
+            LuaMessageUtils.consoleError(scriptName + " Does not exist in " + scriptPath + "! Was there a typo?");
+            return false;
+        }
+        String scriptFilePath = scriptFile.getPath();
 
         String scriptDirectory = LuaCore.getPlugin().getDataFolder() + scriptPath + "/";
         globals.get("package").set("path", globals.get("package").get("path").tojstring() + ";" + scriptDirectory + "?.lua;" + scriptDirectory + "?/?.lua;" + scriptDirectory + "utils/?.lua");
@@ -57,90 +54,63 @@ public class LuaManager {
         // Get the run function from the globals environment
         LuaValue runFunction = globals.get(function);
 
-        // Convert the player object to LuaValue
-        LuaValue playerArg = CoerceJavaToLua.coerce(player);
-        LuaValue eventArgs = CoerceJavaToLua.coerce(event);
-        LuaValue scriptHelper = CoerceJavaToLua.coerce(new ScriptHelper());
 
-        LuaValue[] args = {playerArg, scriptHelper, eventArgs};
+        LuaTable argsTable = new LuaTable();
+        for (LuaArgValue argValue : args) {
+            argsTable.set(argValue.name(), CoerceJavaToLua.coerce(argValue.value()));
+        }
+        argsTable.set("scriptHelper", CoerceJavaToLua.coerce(new ScriptHelper()));
 
-        runFunction.invoke(args);
+
+
+        // Combine all arguments
+        LuaValue[] parsedArgs = {argsTable};
+
+
+        try {
+            runFunction.invoke(parsedArgs);
+        } catch (Exception e) {
+            LuaMessageUtils.consoleError("There was an error while running the script " + scriptName + ".\n" +
+                    "Failed with exception: " + e.getMessage());
+            return false;
+        }
         Long end = System.currentTimeMillis();
         LuaMessageUtils.verboseMessage("Executed " + scriptName + ". The script took " + (end - start) + "ms");
 
         return true;
     }
 
-
     /**
-     * Used for calling a Lua script from Java Code,
-     * In this situation Script path is just the plugins Data Folder
+     * Used for running lua script functions
      * @param function The name of the function you want to run
-     * @param scriptName The name of the Script
-     * @param player Used for parsing in player arguments
-     * @return True if the function exists
+     * @param script The path to the script inclusive (test/test.lua)
+     * @param args The list of argument variables you want to pass in
+     * @return True when all checks are parsed
      */
-    public static boolean runScript(String function, String scriptName, Player player) {
-        return runScript(function, "", scriptName, player, null);
+    public static boolean runScript(String function, String script, LuaArgValue ... args) {
+        return runScript(function,"", script, args);
     }
 
     /**
-     * Used for calling a Lua script from Java Code,
-     * In this situation Script path is just the plugins Data Folder.
-     * This one runs the "run" function
-     * @see LuaManager#runScript(String, String, String, Player, Event) If you want full control
-     * @param scriptName The name of the Script
-     * @param player Used for parsing in player arguments
-     * @return True if the function exists
-     */
-    public static boolean runScript(String scriptName, Player player) {
-
-        return runScript("run", scriptName, player);
-    }
-
-
-    /**
-     * Used for almost full control of running custom functions from the lua
-     * script
-     * @param function The name of the function you want to run
-     * @param path The folder path that the script is in
-     * @param scriptName The name of the script
-     * @param player Used for parsing in player arguments
-     * @return True if the function exists
-     */
-    public static boolean runScript(String function, String path, String scriptName, Player player) {
-        return runScript(function, path, scriptName, player, null);
-    }
-
-
-    /**
-     * Used for running events using the "event" function in the lua script
-     * @param scriptName The name of the script you want to run
-     * @param player Used for parsing in player arguments
-     * @param event The Event that is causing this function to be called
-     *              Used for handling events in your lua script
-     * @return True if the "event" function exists
-     */
-    public static boolean runEvent(String scriptName, Player player, Event event) {
-
-        return runScript("event", "", scriptName, player, event);
-    }
-
-
-    /**
-     * Used for running events using the "event" function in the lua script
+     * Used for running lua script event function
      * @param scriptPath The path to the script
-     * @param scriptName The name of the script you want to run
-     * @param player Used for parsing in player arguments
-     * @param event The Event that is causing this function to be called
-     *              Used for handling events in your lua script
-     * @return True if the "event" function exists
+     * @param scriptName The name of the script file
+     * @param args The list of argument variables you want to pass in
+     * @return True when all checks are parsed
      */
-    public static boolean runEvent(String scriptPath, String scriptName, Player player, Event event) {
-
-        return runScript("event", scriptPath, scriptName, player, event);
+    public static boolean runEvent(String scriptPath, String scriptName, LuaArgValue ... args) {
+        return runScript("event", scriptPath, scriptName, args);
     }
 
+    /**
+     * Used for running lua script functions
+     * @param script The path to the script inclusive (test/test.lua)
+     * @param args The list of argument variables you want to pass in
+     * @return True when all checks are parsed
+     */
+    public static boolean runEvent(String script, LuaArgValue... args) {
+        return runEvent("", script, args);
+    }
     /**
      * Used for checking if a lua function exists without trying to run the lua
      * @param luaScriptPath The path to the script
@@ -192,7 +162,11 @@ public class LuaManager {
 
                 if (file.isFile()) {
                     LuaMessageUtils.verboseMessage("Found file: " + file.getName());
-                    Path destinationPath = new File(LuaCore.getPlugin().getDataFolder() + "/" + folderName, file.getName()).toPath();
+                    File destFile = new File(LuaCore.getPlugin().getDataFolder() + "/" + folderName, file.getName());
+                    if (destFile.exists())
+                        continue;
+                    Path destinationPath = destFile.toPath();
+
                     Path input = new File("/" + file.getParent() + "/" + file.getName()).toPath();
                     destinationPath.toFile().getParentFile().mkdirs();
                     Files.copy(input, destinationPath);
